# api/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"_36_Enums.AttemptStatus": {
				"type": "string",
				"enum": [
					"IN_PROGRESS",
					"SUBMITTED",
					"GRADED"
				]
			},
			"AttemptStatus": {
				"$ref": "#/components/schemas/_36_Enums.AttemptStatus"
			},
			"AttemptSummary": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/AttemptStatus"
					},
					"roomId": {
						"type": "number",
						"format": "double"
					},
					"examId": {
						"type": "number",
						"format": "double"
					},
					"studentId": {
						"type": "number",
						"format": "double"
					},
					"timeTakenSec": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"submittedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"startedAt": {
						"type": "string",
						"format": "date-time"
					},
					"score": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"status",
					"roomId",
					"examId",
					"studentId",
					"timeTakenSec",
					"submittedAt",
					"startedAt",
					"score",
					"id"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Paste the JWT access token obtained from /auth/login"
			}
		}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"description": "A description of QUIZZZ API",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/attempts": {
			"get": {
				"operationId": "ListSubmissions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"total": {
											"type": "number",
											"format": "double"
										},
										"items": {
											"items": {
												"$ref": "#/components/schemas/AttemptSummary"
											},
											"type": "array"
										}
									},
									"required": [
										"total",
										"items"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Attempt"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "examId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateSubmission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttemptSummary"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Attempt"
				],
				"security": [
					{
						"bearerAuth": [
							"STUDENT"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"answers": {
										"items": {
											"properties": {
												"selectedChoiceId": {
													"type": "number",
													"format": "double"
												},
												"questionId": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"selectedChoiceId",
												"questionId"
											],
											"type": "object"
										},
										"type": "array"
									},
									"roomId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"answers",
									"roomId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/attempts/{id}": {
			"get": {
				"operationId": "GetSubmissionById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/AttemptSummary"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Attempt not found",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Attempt"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteSubmission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "number",
											"format": "double"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"id",
										"message"
									],
									"type": "object",
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Attempt not found",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Attempt"
				],
				"security": [
					{
						"bearerAuth": [
							"TEACHER",
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}